Update the existing PopularAuthors widget.
