--- assets/js/modules/analytics-4/components/widgets/MostPopularAuthorsByPageviewsWidget.js
@@ -1,180 +1,0 @@
-/**
- * Site Kit by Google, Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * External dependencies
- */
-import PropTypes from 'prop-types';
-
-/**
- * WordPress dependencies
- */
-import { createInterpolateElement } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
-
-/**
- * Internal dependencies
- */
-import Data from 'googlesitekit-data';
-import { CORE_USER } from '../../../../googlesitekit/datastore/user/constants';
-import {
-	DATE_RANGE_OFFSET,
-	MODULES_ANALYTICS_4,
-} from '../../datastore/constants';
-import {
-	MetricTileTable,
-	MetricTileTablePlainText,
-} from '../../../../components/KeyMetrics';
-import Link from '../../../../components/Link';
-import { ZeroDataMessage } from '../../../analytics/components/common';
-import whenActive from '../../../../util/when-active';
-import ConnectGA4CTATileWidget from './ConnectGA4CTATileWidget';
-import useViewOnly from '../../../../hooks/useViewOnly';
-import { numFmt } from '../../../../util';
-const { useSelect, useInViewSelect } = Data;
-
-function MostPopularAuthorsByPageviewsWidget( props ) {
-	const { Widget } = props;
-
-	const viewOnlyDashboard = useViewOnly();
-
-	const dates = useSelect( ( select ) =>
-		select( CORE_USER ).getDateRangeDates( {
-			offsetDays: DATE_RANGE_OFFSET,
-		} )
-	);
-
-	const reportOptions = {
-		...dates,
-		dimensions: [ 'customEvent:googlesitekit_post_author' ],
-		metrics: [ { name: 'screenPageViews' } ],
-		orderby: [
-			{
-				metric: { metricName: 'screenPageViews' },
-				desc: true,
-			},
-		],
-		limit: 3,
-	};
-
-	const report = useInViewSelect( ( select ) =>
-		select( MODULES_ANALYTICS_4 ).getReport( reportOptions )
-	);
-
-	const error = useSelect( ( select ) =>
-		select( MODULES_ANALYTICS_4 ).getErrorForSelector( 'getReport', [
-			reportOptions,
-		] )
-	);
-
-	const loading = useSelect(
-		( select ) =>
-			! select( MODULES_ANALYTICS_4 ).hasFinishedResolution(
-				'getReport',
-				[ reportOptions ]
-			)
-	);
-
-	global.console.log( report );
-
-	const { rows = [] } = report || {};
-
-	const columns = [
-		{
-			field: 'dimensionValues.0.value',
-			Component: ( { fieldValue } ) => {
-				const url = fieldValue;
-				const title = fieldValue;
-				// Utilizing `useSelect` inside the component rather than
-				// returning its direct value to the `columns` array.
-				// This pattern ensures that the component re-renders correctly based on changes in state,
-				// preventing potential issues with stale or out-of-sync data.
-				// Note: This pattern is replicated in a few other spots within our codebase.
-				const serviceURL = useSelect( ( select ) => {
-					return ! viewOnlyDashboard
-						? select( MODULES_ANALYTICS_4 ).getServiceReportURL(
-								'all-pages-and-screens',
-								{
-									filters: {
-										unifiedPagePathScreen: url,
-									},
-									dates,
-								}
-						  )
-						: null;
-				} );
-
-				if ( viewOnlyDashboard ) {
-					return <MetricTileTablePlainText content={ title } />;
-				}
-
-				return (
-					<Link
-						href={ serviceURL }
-						title={ title }
-						external
-						hideExternalIndicator
-					>
-						{ title }
-					</Link>
-				);
-			},
-		},
-		{
-			field: 'metricValues.0.value',
-			Component: ( { fieldValue } ) =>
-				createInterpolateElement( '<metricValue /> CTR', {
-					metricValue: (
-						<strong>
-							{ numFmt( fieldValue, {
-								style: 'percent',
-								maximumFractionDigits: 1,
-							} ) }
-						</strong>
-					),
-				} ),
-		},
-	];
-
-	return (
-		<MetricTileTable
-			Widget={ Widget }
-			title={ __(
-				'Most popular authors by pageviews',
-				'google-site-kit'
-			) }
-			loading={ loading }
-			rows={ rows }
-			columns={ columns }
-			ZeroState={ ZeroDataMessage }
-			error={ error }
-			moduleSlug="analytics-4"
-			infoTooltip={ __(
-				'Authors whose posts got the most visits',
-				'google-site-kit'
-			) }
-		/>
-	);
-}
-
-MostPopularAuthorsByPageviewsWidget.propTypes = {
-	Widget: PropTypes.elementType.isRequired,
-};
-
-export default whenActive( {
-	moduleName: 'analytics-4',
-	FallbackComponent: ConnectGA4CTATileWidget,
-} )( MostPopularAuthorsByPageviewsWidget );
