+++ assets/js/modules/analytics-4/components/widgets/PopularAuthorsWidget.js
@@ -25,24 +25,154 @@ import PropTypes from 'prop-types';
  * WordPress dependencies
  */
 import { compose } from '@wordpress/compose';
+import { createInterpolateElement } from '@wordpress/element';
+import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import { KM_ANALYTICS_POPULAR_AUTHORS } from '../../../../googlesitekit/datastore/user/constants';
-import { MetricTileTable } from '../../../../components/KeyMetrics';
+import Data from 'googlesitekit-data';
+import {
+	CORE_USER,
+	KM_ANALYTICS_POPULAR_AUTHORS,
+} from '../../../../googlesitekit/datastore/user/constants';
+import {
+	DATE_RANGE_OFFSET,
+	MODULES_ANALYTICS_4,
+} from '../../datastore/constants';
+import {
+	MetricTileTable,
+	MetricTileTablePlainText,
+} from '../../../../components/KeyMetrics';
+import Link from '../../../../components/Link';
 import { ZeroDataMessage } from '../../../analytics/components/common';
-import whenActive from '../../../../util/when-active';
 import withCustomDimensions from '../../utils/withCustomDimensions';
+import whenActive from '../../../../util/when-active';
 import ConnectGA4CTATileWidget from './ConnectGA4CTATileWidget';
+import useViewOnly from '../../../../hooks/useViewOnly';
+import { numFmt } from '../../../../util';
+const { useSelect, useInViewSelect } = Data;
 
-function PopularAuthorsWidget( { Widget } ) {
+function getPopularAuthorsWidgetReportOptions( select ) {
+	const dates = select( CORE_USER ).getDateRangeDates( {
+		offsetDays: DATE_RANGE_OFFSET,
+	} );
+
+	return {
+		...dates,
+		dimensions: [ 'customEvent:googlesitekit_post_author' ],
+		metrics: [ { name: 'screenPageViews' } ],
+		orderby: [
+			{
+				metric: { metricName: 'screenPageViews' },
+				desc: true,
+			},
+		],
+		limit: 3,
+	};
+}
+
+function PopularAuthorsWidget( props ) {
+	const { Widget } = props;
+
+	const viewOnlyDashboard = useViewOnly();
+
+	const reportOptions = useSelect( getPopularAuthorsWidgetReportOptions );
+	const report = useInViewSelect( ( select ) =>
+		select( MODULES_ANALYTICS_4 ).getReport( reportOptions )
+	);
+
+	const error = useSelect( ( select ) =>
+		select( MODULES_ANALYTICS_4 ).getErrorForSelector( 'getReport', [
+			reportOptions,
+		] )
+	);
+
+	const loading = useSelect(
+		( select ) =>
+			! select( MODULES_ANALYTICS_4 ).hasFinishedResolution(
+				'getReport',
+				[ reportOptions ]
+			)
+	);
+
+	const { rows = [] } = report || {};
+
+	const columns = [
+		{
+			field: 'dimensionValues.0.value',
+			Component: ( { fieldValue } ) => {
+				const url = fieldValue;
+				const title = fieldValue;
+				// Utilizing `useSelect` inside the component rather than
+				// returning its direct value to the `columns` array.
+				// This pattern ensures that the component re-renders correctly based on changes in state,
+				// preventing potential issues with stale or out-of-sync data.
+				// Note: This pattern is replicated in a few other spots within our codebase.
+				const serviceURL = useSelect( ( select ) => {
+					return ! viewOnlyDashboard
+						? select( MODULES_ANALYTICS_4 ).getServiceReportURL(
+								'all-pages-and-screens',
+								{
+									filters: {
+										unifiedPagePathScreen: url,
+									},
+									dates: {},
+								}
+						  )
+						: null;
+				} );
+
+				if ( viewOnlyDashboard ) {
+					return <MetricTileTablePlainText content={ title } />;
+				}
+
+				return (
+					<Link
+						href={ serviceURL }
+						title={ title }
+						external
+						hideExternalIndicator
+					>
+						{ title }
+					</Link>
+				);
+			},
+		},
+		{
+			field: 'metricValues.0.value',
+			Component: ( { fieldValue } ) =>
+				createInterpolateElement( '<metricValue /> CTR', {
+					metricValue: (
+						<strong>
+							{ numFmt( fieldValue, {
+								style: 'percent',
+								maximumFractionDigits: 1,
+							} ) }
+						</strong>
+					),
+				} ),
+		},
+	];
+
 	return (
 		<MetricTileTable
 			widgetSlug={ KM_ANALYTICS_POPULAR_AUTHORS }
 			Widget={ Widget }
+			title={ __(
+				'Most popular authors by pageviews',
+				'google-site-kit'
+			) }
+			loading={ loading }
+			rows={ rows }
+			columns={ columns }
 			ZeroState={ ZeroDataMessage }
+			error={ error }
 			moduleSlug="analytics-4"
+			infoTooltip={ __(
+				'Authors whose posts got the most visits',
+				'google-site-kit'
+			) }
 		/>
 	);
 }
@@ -56,5 +186,7 @@ export default compose(
 		moduleName: 'analytics-4',
 		FallbackComponent: ConnectGA4CTATileWidget,
 	} ),
-	withCustomDimensions()
+	withCustomDimensions( {
+		reportOptions: getPopularAuthorsWidgetReportOptions,
+	} )
 )( PopularAuthorsWidget );
